graveyard repository definition
