+++ DiscordBot.Data/DiscordBot.Data.csproj
@@ -7,14 +7,14 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\DiscordBot.Common\DiscordBot.Common.csproj"/>
+        <ProjectReference Include="..\DiscordBot.Common\DiscordBot.Common.csproj" />
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="LiteDB" Version="5.0.11"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.0"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.0.0"/>
-        <PackageReference Include="Serilog" Version="2.10.0"/>
+        <PackageReference Include="LiteDB" Version="5.0.11" />
+        <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.0.0" />
+        <PackageReference Include="Serilog" Version="2.10.0" />
     </ItemGroup>
 
 </Project>
